Vocard-Fresh/requirements.txt
# Vocard Bot - Dependencies for Railway Deployment
discord.py==2.5.2
motor==3.6.0
dnspython==2.6.1
tldextract==3.2.1
validators==0.18.2
humanize==4.0.0
beautifulsoup4==4.11.1
psutil==5.9.8
aiohttp>=3.11.0
python-dotenv==1.1.1
lyricsgenius==3.0.1
geoip2==4.8.1
PyYAML>=6.0
websockets>=12.0
PyNaCl>=1.5.0
yt-dlp>=2024.0.0
Pillow>=10.0.0
aiofiles>=23.0.0
akinator>=2.0.2
curl_cffi>=0.14.0